Move openai chat client functionality to its LLM, remove separate Chat
The chat llm client calls use the newer OpenAI API to implement the
same functionality
