+++ llms/openai/multicontent_test.go
@@ -13,21 +13,21 @@ import (
 	"github.com/tmc/langchaingo/schema"
 )
 
-func newChatClient(t *testing.T, opts ...Option) *Chat {
+func newTestClient(t *testing.T, opts ...Option) *LLM {
 	t.Helper()
 	if openaiKey := os.Getenv("OPENAI_API_KEY"); openaiKey == "" {
 		t.Skip("OPENAI_API_KEY not set")
 		return nil
 	}
 
-	llm, err := NewChat(opts...)
+	llm, err := New(opts...)
 	require.NoError(t, err)
 	return llm
 }
 
 func TestMultiContentText(t *testing.T) {
 	t.Parallel()
-	llm := newChatClient(t)
+	llm := newTestClient(t)
 
 	parts := []llms.ContentPart{
 		llms.TextContent{Text: "I'm a pomeranian"},
@@ -50,7 +50,7 @@ func TestMultiContentText(t *testing.T) {
 
 func TestMultiContentTextChatSequence(t *testing.T) {
 	t.Parallel()
-	llm := newChatClient(t)
+	llm := newTestClient(t)
 
 	content := []llms.MessageContent{
 		{
@@ -78,7 +78,7 @@ func TestMultiContentTextChatSequence(t *testing.T) {
 func TestMultiContentImage(t *testing.T) {
 	t.Parallel()
 
-	llm := newChatClient(t, WithModel("gpt-4-vision-preview"))
+	llm := newTestClient(t, WithModel("gpt-4-vision-preview"))
 
 	parts := []llms.ContentPart{
 		llms.ImageURLContent{URL: "https://github.com/tmc/langchaingo/blob/main/docs/static/img/parrot-icon.png?raw=true"},
@@ -101,7 +101,7 @@ func TestMultiContentImage(t *testing.T) {
 
 func TestWithStreaming(t *testing.T) {
 	t.Parallel()
-	llm := newChatClient(t)
+	llm := newTestClient(t)
 
 	parts := []llms.ContentPart{
 		llms.TextContent{Text: "I'm a pomeranian"},
@@ -132,7 +132,7 @@ func TestWithStreaming(t *testing.T) {
 //nolint:lll
 func TestFunctionCall(t *testing.T) {
 	t.Parallel()
-	llm := newChatClient(t)
+	llm := newTestClient(t)
 
 	parts := []llms.ContentPart{
 		llms.TextContent{Text: "What is the weather like in Boston?"},
